// src/gen/models/postApiTransactionsDepositsBody.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Tasks API
 * OpenAPI spec version: 0.0.0
 */

export type PostApiTransactionsDepositsBody = {
  id: string;
  /** @nullable */
  processedAt: string | null;
  /** @nullable */
  walletId: string | null;
  type: string;
  status: string;
  amount: number;
  /** @nullable */
  netAmount: number | null;
  /** @nullable */
  currencyName: string | null;
  /** @nullable */
  feeAmount: number | null;
  /** @nullable */
  productId: string | null;
  /** @nullable */
  paymentMethod: string | null;
  /** @nullable */
  balanceBefore: number | null;
  /** @nullable */
  balanceAfter: number | null;
  /** @nullable */
  bonusBalanceBefore: number | null;
  /** @nullable */
  bonusBalanceAfter: number | null;
  /** @nullable */
  bonusAmount: number | null;
  /** @nullable */
  wageringRequirement: number | null;
  /** @nullable */
  wageringProgress: number | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  provider: string | null;
  /** @nullable */
  providerTxId: string | null;
  /** @nullable */
  relatedGameId: string | null;
  /** @nullable */
  relatedRoundId: string | null;
  /** @nullable */
  createdAt: string | null;
  /** @nullable */
  updatedAt: string | null;
  /** @nullable */
  operatorId: string | null;
  userId: string;
};